add targeting/search overlay
it will create an overlay that appears on the screen everytime i click
on soemthing on the screen after gameStart. this will mantain its aim on
the coordinates. needs to add clickable options as well as update the coordinates on firebase or at least see if i can go back to the old system