add Chameleon, restricts view of final score to 5 items, sorts scores to best, remove scores at 0
final commit in terms of functionality, from now on is all style adjustment and minor
styling. Refactoring of states will follow to make sure that they are eaily readable
(eg: isGameOver instead of gameOver, to follow the convention on boolean variables)
the scoresOverlay has been adjusted so that it only shows 5 scores which are sorted
from smaller to bigger. an additional div shows the current score of the player. the
Chameleon item has been added as well, bringing the final count of items to be found
to 3.