refactor AlertOverlay state in SearchOverlay component and ui state in Main component
all the boolean variables are now written with readability in mind, such as
"isGameOver" instead of "gameOver" as well as "isInstructionOverlayOpen" instead of
"instruction". Same for AlertOverlay state where "show" is now "isAlertShown". Next, i should refactor the function that search the coordinates to be a single finction isntead of 3