adds function find(name) instead of destructuring into 3 functions for each dataset
this happens by using a switch case to determine which set of data to load on the
function depending on the name passed to it. previously attempts were made by using comouted properties for vanilla js which are not supported in react, so a switch was implemented instead. this way i removed 3 functions